requirements.txt
python-telegram-bot==20.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
openai==1.3.0
nest_asyncio==1.5.8